requirements.txt
# YOLOv5 GradCAM++ and Cut-and-Paste Validation Requirements

# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0

# YOLOv5 dependencies
matplotlib>=3.3.0
pandas>=1.1.0
seaborn>=0.11.0
tqdm>=4.62.0
pyyaml>=5.4.0

# Image processing
scikit-image>=0.18.0
albumentations>=1.0.0

# Statistical analysis
scipy>=1.7.0
statsmodels>=0.12.0

# Visualization
plotly>=5.0.0
notebook>=6.4.0
ipywidgets>=7.6.0

# Utilities
requests>=2.26.0
psutil>=5.8.0
tensorboard>=2.7.0

# Optional GPU support
# torch-cuda>=11.3  # Uncomment if using GPU